Fix declaration of qHash(TableArea, ...)
git-svn-id: svn://scribus.net/trunk/Scribus@27620 11d20701-8431-0410-a711-e3c959e3b870

+++ scribus/styles/tablearea.h
@@ -35,7 +35,7 @@ enum class TableArea
 
 Q_DECLARE_METATYPE(TableArea)
 
-inline uint qHash(TableArea area, uint seed = 0)
+inline size_t qHash(TableArea area, size_t seed = 0)
 {
 	return ::qHash(static_cast<int>(area), seed);
 }
